Add a potentiometer for speed control
Add a potentiometer at analog pin A0 for speed control of the chase sequence.  A variable is created within the for loop that reads the analog voltage fo the pot and maps it output from 10 - 1000 (in milliseconds) into a new variable called delayTime which is then used to set the delay time of the chase.