add a while loop for better speed control
added a variable = v and a while loop to control the speed of the program.  a variable called v is tested in a while loop then incremented so that the while loop only runs once.  at the end of the program the variable v is set to 0 again so that next time through the loop, the while loop will run again